Fix non deterministic test behavior
When executing the network to return the answer for a given image data, the
results varied in a small margin. Debugging the code we discovered that this
issue was caused by tf.train.string_input_producer, tf.train.shuffle_batch and
tf.train.batch, which all use threads internally to increase the speed of
computation. The use of threads was chaging the order of the input data given.
The solution was to use only one thread when evaluating (or testing) the
network.

A similar issue can be found here:
https://github.com/tensorflow/tensorflow/issues/410